requirement.txt
APScheduler==3.6.3
cachelib==0.6.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
dnspython==2.2.1
Flask==2.0.3
Flask-Session==0.4.0
google-api-core==2.6.0
google-api-python-client==2.39.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
oauthlib==3.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.0.2
pyparsing==3.0.7
python-decouple==3.6
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-oauthlib==1.3.1
rfc3339==6.2
rsa==4.8
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
